Just added some more comments

+++ src/com/Emile2250/SimpleUHC/SimpleUHC.java
@@ -1,15 +1,12 @@
 package com.Emile2250.SimpleUHC;
 
-import com.Emile2250.SimpleUHC.Commands.AdminCommands;
 import com.Emile2250.SimpleUHC.Commands.CommandHandler;
-import com.Emile2250.SimpleUHC.Commands.PlayerCommands;
 import com.Emile2250.SimpleUHC.Listeners.LeaveEvent;
 import com.Emile2250.SimpleUHC.Listeners.UHCDeath;
 import com.Emile2250.SimpleUHC.Listeners.UHCPvp;
 import com.Emile2250.SimpleUHC.UHC.Game;
 import org.apache.commons.io.FileUtils;
 import org.bukkit.Bukkit;
-import org.bukkit.World;
 import org.bukkit.configuration.InvalidConfigurationException;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.configuration.file.YamlConfiguration;
@@ -18,7 +15,6 @@ import org.bukkit.plugin.java.JavaPlugin;
 import java.io.File;
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.List;
 
 public class SimpleUHC extends JavaPlugin {
 
@@ -90,18 +86,19 @@ public class SimpleUHC extends JavaPlugin {
     }
 
     private void deleteWorlds() {
-        String[] directories = this.getServer().getWorldContainer().list();
+        String[] directories = this.getServer().getWorldContainer().list(); // Grabs a list of all files in the server
 
-        for (String folder : directories) {
-            if (folder.contains("UHC-")) {
-                try {
-                    FileUtils.deleteDirectory(new File(this.getServer().getWorldContainer(), folder));
-                } catch (IOException e) {
-                    System.out.println("Oh no! We had an issue deleting left over worlds");
-                    e.printStackTrace();
+        if (directories != null) {
+            for (String folder : directories) { // Loops through all the files
+                if (folder.contains("UHC-")) { // Checks if we had any left over UHC crashes in case of shutdown mid game
+                    try {
+                        FileUtils.deleteDirectory(new File(this.getServer().getWorldContainer(), folder)); // Tries to delete the world.
+                    } catch (IOException e) {
+                        System.out.println("Oh no! We had an issue deleting left over worlds");
+                        e.printStackTrace();
+                    }
                 }
             }
         }
     }
-
 }
